refactor: split submit flow into modules
